Move out arithmetic ops
